fix div . change variables data type int > float handle commite warning

+++ app/src/main/java/education/mostafa/projects/servicekt/interfaces/main_view.kt
@@ -4,5 +4,5 @@ interface main_view {
     fun initViews()
     fun initOperations()
     fun validInputs():Boolean
-    fun setRest(res:Int)
+    fun setRest(res:Float)
 }
